+++ Isolator.Fody/AssemblyLoader/MsCoreReferenceFinder.cs
@@ -22,4 +22,13 @@ public partial class ModuleWeaver
         var compilerGeneratedAttribute = FindTypeDefinition("System.Runtime.CompilerServices.CompilerGeneratedAttribute");
         _compilerGeneratedAttributeCtor = ModuleDefinition.ImportReference(compilerGeneratedAttribute.Methods.First(_ => _.IsConstructor));
     }
+
+    internal void AddCompilerGeneratedAttribute(ICustomAttributeProvider customAttributeProvider)
+    {
+        if (customAttributeProvider.CustomAttributes.Any(_ => _.AttributeType.FullName == "System.Runtime.CompilerServices.CompilerGeneratedAttribute"))
+        {
+            return;
+        }
+        customAttributeProvider.CustomAttributes.Add(new CustomAttribute(_compilerGeneratedAttributeCtor));
+    }
 }