+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `IsolatorAssemblyLoadContext` to support multiple resolvers and search in all resolvers.
 - Update `Isolator.ConsoleApp` and remove random class.
 - Update clone method to use `CompilerGenerated` attribute.
+- Update to add `CompilerGenerated` attribute in the methods and type using `AddCompilerGeneratedAttribute`.
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0